% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete-wp-posts.R
\name{delete_wp_post}
\alias{delete_wp_post}
\title{Delete a WordPress Post}
\usage{
delete_wp_post(root_url, user, pass, post_id, verbose = TRUE)
}
\arguments{
\item{root_url}{The domain on which you wish to create the post.}

\item{user}{The username to be passed into the API call to create the post.}

\item{pass}{The password to be used in the API call to create the post.
To get this value, you must have the Application Passwords plugin
installed, and must create an application using that plugin via your
WordPress admin panel; there you will get the password needed.}

\item{post_id}{The text string associated with the tag.}

\item{verbose}{If TRUE, the tag id will be printed to the console. Potentially useful
for usage in a single-threaded loop.}
}
\value{
response from the API as a list object. Look for the key "deleted" within the list
to confirm the post was deleted.
}
\description{
Delete a post on your WordPress site using your WordPress site's API.
}
\examples{
\dontrun{
delete_wp_tag(root_url = 'https://domain.com',user = Sys.getenv('username'),
pass = Sys.getenv('password'),post_id = 12)
}

}
