% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-wp-authors.R
\name{get_wp_authors}
\alias{get_wp_authors}
\title{Get WordPress Authors}
\usage{
get_wp_authors(root_url)
}
\arguments{
\item{root_url}{The WordPress site for which the data is sought.}
}
\value{
A dataframe with four values: the user ID for the author in the site's WordPress database; the author's URL on the site; the author's name in the database; and the value of the description field associated with the author.

#'\dontrun{
get_wp_authors('domain.com')
}
}
\description{
Retrieves a list of all publishers on the site.
}
