% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pods.R
\name{get_wp_catlike_pod}
\alias{get_wp_catlike_pod}
\title{Get a list of Category-like Pod}
\usage{
get_wp_catlike_pod(route)
}
\description{
If you're using the Pods plugin and have created
a type of pod that is essentially like a category, this
function will help you get all the "categories" in a given route.
You'll need to have an environment variable called
'wp_domain' that contains the base
URL of your site (i.e. https://example.com).
}
