% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-wp-pages.R
\name{get_wp_pages}
\alias{get_wp_pages}
\title{Retrieve WordPress Pages}
\usage{
get_wp_pages(root_url, page_count = Inf, after_date = NULL)
}
\arguments{
\item{root_url}{The WordPress site for which pages are sought to be retrieved.}

\item{after_date}{The date after which pages should be returned.}

\item{post_count}{The maximum number of pages to return, sorted by the most recent date.
Default value is to return all pages made on the site.}
}
\value{
A data frame returning the post ID, publication date, title, excerpt, content,
tag IDs, category IDs, and author IDs.
}
\description{
Retrieve pages made on the WordPress site.
}
\examples{
\dontrun{
get_wp_pages(root_url = 'https://domain.com',page_count = 200, after_date = NULL)
}

}
