% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-wp-posts-after.R
\name{get_wp_posts_after}
\alias{get_wp_posts_after}
\title{Retrieve WordPress Posts After a Certain Date}
\usage{
get_wp_posts_after(root_url, after_date)
}
\arguments{
\item{root_url}{The WordPress site for which posts are sought to be retrieved.}

\item{after_date}{The date after which posts should be returned. Should be
in "YYYY-MM-DD" format}
}
\value{
A data frame returning the post ID, publication date, title, excerpt,
content, tag IDs,category IDs, and author IDs.
}
\description{
Retrieve posts made on the WordPress site.
}
\examples{
\dontrun{
get_wp_posts(root_url = 'https://domain.com',post_count = 200, after_date = NULL)
}

}
