% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tags.R
\name{get_wp_tags}
\alias{get_wp_tags}
\title{Get WordPress Tags}
\usage{
get_wp_tags(root_url)
}
\arguments{
\item{root_url}{The WordPress site for which the data is sought.}
}
\value{
A dataframe with three values: the slug for a given tag, its ID in the WordPress site database, and the human readable tag name.
}
\description{
Retrieves a list of all tags created on a given WordPress site.
}
\examples{
#'\dontrun{
get_wp_tags('domain.com')
}
}
