% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-wp-users.R
\name{get_wp_users}
\alias{get_wp_users}
\title{Retrieve WordPress User List}
\usage{
get_wp_users(root_url)
}
\arguments{
\item{root_url}{The domain for which users are sought to be retrieved.}
}
\value{
A dataframe with three columns: the ID of the user in the WordPress database; their user name on the site; and the URL of their user page on the site.
}
\description{
Returns all the registered users on a given WordPress site that can be retrieved without a key.
}
\examples{
\dontrun{
get_wp_users(root_url = 'domain.com')
}

}
