% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pods.R
\name{update_wp_catlike_pod}
\alias{update_wp_catlike_pod}
\title{Update a Category-like Pod}
\usage{
update_wp_catlike_pod(route, id, name, slug, description)
}
\description{
If you're using the Pods plugin and have created
a type of pod that is essentially like a category, this
function will help you work with that endpoint. You'll need to have
an environment variable called 'wp_domain' that contains the base
URL of your site (i.e. https://example.com).
}
