% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-wp-cats.R
\name{get_wp_cats}
\alias{get_wp_cats}
\title{Get WordPress Categories}
\usage{
get_wp_cats(root_url)
}
\arguments{
\item{root_url}{}
}
\value{
A data frame, unique by category ID, with 7 columns: category ID, category name, category slug, post count, description, URL, and parent category.

\dontrun{
get_wp_cats(root_url = 'domain.com')
}
}
\description{
Retrieves the categories created on a WordPress site.
}
